Split motor controls into a separte file. Added if statements for other inputs. Pseudocode for boost.
